Fix circular dependencies. Fix webpack build (#174)
