+++ src/credentials/index.ts
@@ -4,7 +4,6 @@ export * from './status/sparse-merkle-tree';
 export * from './status/resolver';
 export * from './status/agent-revocation';
 export * from './status/credential-status-publisher';
-export * from './status/utils';
 export * from './credential-wallet';
 export * from './rhs';
 export * from './utils';
